refactor: move uploaded-font loading into the uploadedFonts module
_loadFormPackages had the whole registration block inline: registry
population, alias computation, FontFace creation, and the CloudFront-to-S3
retry. All of it is uploaded-font identity logic, so it belongs beside the
alias it depends on — and it leaves featheryClient with a single call,
mirroring how Google fonts are loaded.

The region-aware S3 host is passed in rather than imported, since
featheryClient owns that value and importing it here would close a cycle.

Side benefit: this drops the diff in featheryClient/index.ts to two lines,
so it no longer collides with #1754, which rewrites the adjacent
Google-font block.
