+++ src/utils/featheryClient/index.ts
@@ -23,11 +23,7 @@ import { loadPhoneValidator } from '../validation';
 import { initializeIntegrations } from '../../integrations/utils';
 import { loadLottieLight } from '../../elements/components/Lottie';
 import { downloadAllFileUrls, featheryDoc, featheryWindow } from '../browser';
-import {
-  isAliasableUploadName,
-  registerUploadedFonts,
-  uploadedFontRenderFamily
-} from '../uploadedFonts';
+import { loadUploadedFonts } from '../uploadedFonts';
 import { authState } from '../../auth/LoginForm';
 import { loadQRScanner } from '../../elements/fields/QRScanner/qrLoader';
 import { gatherTrustedFormFields } from '../../integrations/trustedform';
@@ -399,30 +395,8 @@ export default class FeatheryClient extends IntegrationClient {
         WebFont.load({ google: { families: res.fonts } });
       });
     }
-    // Load user-uploaded fonts. Registered under an aliased family so an
-    // upload sharing a catalog font's name can't shadow it — style values are
-    // rewritten to match in transformFontFamilies, which reads this registry.
-    registerUploadedFonts(Object.keys(res.uploaded_fonts));
-    Object.entries(res.uploaded_fonts).forEach(([family, fontStyles]) => {
-      const renderFamily = isAliasableUploadName(family)
-        ? uploadedFontRenderFamily(family)
-        : family;
-      (fontStyles as any).forEach(({ source, style, weight }: any) => {
-        const loadFont = (url: string) =>
-          new FontFace(renderFamily, `url(${url})`, { style, weight })
-            .load()
-            .then((font) => featheryDoc().fonts.add(font));
-        loadFont(source).catch(() => {
-          // Cloudfront might run into CORS issues so fall back to
-          // S3 directly if needed
-          const fallback = new URL(source);
-          fallback.hostname = S3_URL;
-          loadFont(fallback.toString()).catch((e) =>
-            console.warn(`Font load issue: ${e}`)
-          );
-        });
-      });
-    });
+    // Load user-uploaded fonts
+    loadUploadedFonts(res.uploaded_fonts, S3_URL);
     // Load Lottie if form needs animations
     let needLottie = false;
     // Load phone number validator for phone and login fields
